+++ src/selectors/articles/select_article.ts
@@ -4,7 +4,7 @@ import { selectLocale } from '../locale/select_locale';
 import { selectRelatedTasks } from '../tasks';
 import { RouterProps } from '../../application/routing';
 import { selectExploreTaxonomy } from '../taxonomies';
-import { selectIconFromExploreTaxonomy } from '../select_icon_from_explore_taxonomy';
+import { selectIconFromExploreTaxonomy } from '../explore/select_icon_from_explore_taxonomy';
 import { toSelectorArticle } from './to_selector_article';
 import { toSelectorArticleList } from './to_selector_article_list';
 import { Article } from './types';
